ENdilich fertig

+++ World Weather/World Weather/Models/DashboardModel.cs
@@ -8,10 +8,10 @@ namespace World_Weather.Models
 {
     public class DashboardModel
     {
-        public int Wien { get; set; }
-        public int NewYork { get; set; }
-        public int Sydney { get; set; }
-        public int Tokio { get; set; }
+        public double Wien { get; set; }
+        public double NewYork { get; set; }
+        public double Sydney { get; set; }
+        public double Tokio { get; set; }
         public List<string> CityPickOptions { get; set; }
 
         public string SelectedCapital { get; set; }
